refactor sidebar to use base component
